fix: move lib/sms to utils/server/sms
