+++ src/app/api/public/sms/events/fails/route.ts
@@ -3,9 +3,8 @@ import 'server-only'
 import * as Sentry from '@sentry/nextjs'
 import { NextRequest, NextResponse } from 'next/server'
 
+import { verifySignature } from '@/utils/server/sms'
 import { getLogger } from '@/utils/shared/logger'
-
-import { verifySignature } from '@/lib/sms'
 
 interface SmsEvent {
   ParentAccountSid: string
